+++ app/views/layout/default.html.php
@@ -25,58 +25,13 @@
   </head>
   <body>
 
-    <header>
-      <h1>
-        <?php if (fz_config_get ('looknfeel', 'your_logo', '') != ''): ?>
-          <span id="your-logo">
-            <img src="<?php echo public_url_for (fz_config_get ('looknfeel', 'your_logo')) ?>"/>
-          </span>
-        <?php endif ?>
-        <span id="filez-header">
-          <a href="<?php echo public_url_for ('/') ?>" id="filez-logo">
-            <img src="<?php echo public_url_for ('resources/images/filez-logo.png') ?>" title="filez" />
-          </a>
-          <?php echo __('Share files for a limited time.') ?>
-        </span>
-        <span style="display: block; clear: both;"></span>
-      </h1>
-      <?php if (array_key_exists ('notification', $flash)): ?>
-        <p class="notif ok"><?php echo $flash ['notification'] ?></p>
-      <?php endif ?>
-      <?php if (array_key_exists ('error', $flash)): ?>
-        <p class="notif error"><?php echo $flash ['error'] ?></p>
-      <?php endif ?>
-
-      <?php if (isset ($user)): ?>
-        <p id="auth-box">
-          <?php echo $user ['email'] ?> |
-          <a href="<?php echo url_for ('/logout') ?>" title="<?php echo __('Log out') ?>">&nbsp;</a>
-        </p>
-      <?php endif ?>
-    </header>
+    <?php echo partial ('layout/_header.php', $vars); ?>
 
     <article>
       <?php echo $content ?>
     </article>
 
-    <footer>
-      <?php if (is_array (option ('debug_msg'))): ?>
-        <div class="debug"><h3>Logged messages :</h3>
-        <?php foreach (option ('debug_msg') as $msg): ?>
-          <pre><?php echo $msg ?></pre>
-        <?php endforeach ?>
-        </div>
-      <?php endif ?>
-
-      <?php if (isset ($user)): ?>
-        <p id="disk-usage"><?php echo __r('Using %space% of %quota%', array (
-            // TODO this code should not be here
-            'space' => '<b>'.bytesToShorthand (Fz_Db::getTable('File')->getTotalDiskSpaceByUser ($user)).'</b>',
-            'quota' => fz_config_get('app', 'user_quota'))); ?>.
-        </p>
-      <?php endif ?>
-      <a href="http://gpl.univ-avignon.fr">Un logiciel libre de l'Université d'Avignon et des Pays de Vaucluse</a>
-    </footer>
+    <?php echo partial ('layout/_footer.php', $vars); ?>
 
     <div id="modal-background"></div>
   </body>