+++ app/views/layout/error.html.php
@@ -15,30 +15,11 @@
     <!--[if lte IE 8]>
     <script type="text/javascript" src="<?php echo public_url_for ('resources/js/html5.js') ?>"></script>
     <![endif]-->
-    <header>
-      <h1>
-        <?php if (fz_config_get ('looknfeel', 'your_logo', '') != ''): ?>
-          <span id="your-logo">
-            <img src="<?php echo public_url_for (fz_config_get ('looknfeel', 'your_logo')) ?>"/>
-          </span>
-        <?php endif ?>
-        <span id="filez-header">
-          <a href="<?php echo public_url_for ('/') ?>" id="filez-logo">
-            <img src="<?php echo public_url_for ('resources/images/filez-logo.png') ?>" title="filez" />
-          </a>
-          <?php echo __('Share files for a limited time.') ?>
-        </span>
-        <span style="display: block; clear: both;"></span>
-      </h1>
-        
-      <?php if (array_key_exists ('notification', $flash)): ?>
-        <p class="notif ok"><?php echo $flash ['notification'] ?></p>
-      <?php endif ?>
-      <?php if (array_key_exists ('error', $flash)): ?>
-        <p class="notif error"><?php echo $flash ['error'] ?></p>
-      <?php endif ?>
 
-    </header>
+  </head>
+  <body>
+
+    <?php echo partial ('layout/_header.php', $vars); ?>
 
     <article>
       <div class="notif error">
@@ -46,9 +27,7 @@
       </div>
     </article>
 
-    <footer>
-      <a href="http://gpl.univ-avignon.fr">Un logiciel libre de l'Université d'Avignon et des Pays de Vaucluse</a>
-    </footer>
+    <?php echo partial ('layout/_footer.php', $vars); ?>
 
   </body>
 </html>